The Books Update
This change adds 6e Translation Master Document, 4e Translation Master Document, and the Master Player Rules Document as valid books, and changes what I believe to be most if not all instances of items from these documents to specify which one they come from.

The goal here being to make it easier to find the source.

It also fixes a problem with the Cryo Minigrenade I happened to find in amend_weapons.xml which in all honesty should be its own fix but I was already working on this soooooooooo