Changed the amend-lifestyles.xml
Added Illness Medication and Antidepressants as a quality for Lifestyles. Have confirmed that this also is not impacted by multiplicative qualities like Sterile Sanctuary.

Also fixed a potential issue with Sterile Sanctuary and Uncontrolled Metastasis which had a typo on Category that seemed to have prevented it from showing up.